refactor(generator): extract EndpointChecks, trim DiagnosticEmitter to orchestrator
